to_string tests for Cards and Hands (#39)
* Added tests for to_string methods

* fixed formatting, added partially filled hand test

* Stylistic and formatting changes

Some things I fixed:
- 4 spaces for indentation 
- Spaces after `(`, `[`, `<`, `{` (array/object initializer braces) and before `)`, `]`, `>`, and `}` unless there's nothing between them
- No more than 1 consecutive blank line, if you're sectioning things off, put a explanatory block comment
- Use `<` and `>` for `#including` files from another directory. In the case for tests, this is everything but `catch.hpp`

Stylistically, I think if you're going to split strings across new lines with `\n`s in them, you should do it on the `\n` characters themselves. Personally, I find it much easier to visualize what the text *should* look like if you do this

* Changed test case descriptions

A section should describe exactly what is being tested in human language. This should be unambiguous but also brief, kind of like a requirement.

When there are a small, finite number of conditions, it's probably better to bundle them all into a single test case/section just to be additionally thorough, even though it's not strictly necessary, unless the code being tested is extremely critical.

* Improved test sanitation

Try to keep the variables confined to only the scope absolutely necessary. If a test case has sections, then each section should be *completely* independent of one another so that you can't possibly have one section interfering with another.

Another good tip is to eliminate the hard-coded strings with all the repetition. If the value in the loop is being specifically assigned to something, then you can also build the expected string to be what you would expect that card to be serialized to. It just makes it a bit easier and more adaptive to changes (although, some might argue that tests *shouldn't* be that way, and I can see their viewpoint as well).

The face-down tests should also make sure that the second card is treated normally though.

* Removed unused variables

this one was entirely on me -- oops!